Moved var-cov into own function.
